feat: completed the landing page
